Refaktoroi: Siirrä rouhintalogiikkaa laajoihin tyyppeihin;
Tämä on askel sitä kohti, että päästään käyttämään laajojen tyyppien kautta rouhinnan
tuloksissa myös tietoja kuntailmoituksista.

Parhaillaan suorittaminen tutkitaan nyt vähän yksinkertaisemmin, koska siihen riittää, että joitain
opintoja ollaan suorittamassa: viimeisimpien oppivelvollisuutietojen mukaan järjestäminen oli turha
vaihe.
